extracted bitbuffer code from ojph_precinct.cpp
