+++ src/core/codestream/ojph_precinct.cpp
@@ -42,11 +42,12 @@
 #include "ojph_file.h"
 #include "ojph_mem.h"
 #include "ojph_params.h"
-#include "ojph_codestream.h"
 #include "ojph_codestream_local.h"
 #include "ojph_precinct.h"
 #include "ojph_subband.h"
 #include "ojph_codeblock.h" // for coded_cb_header
+#include "ojph_bitbuffer_write.h"
+#include "ojph_bitbuffer_read.h"
 
 #include "../transform/ojph_colour.h"
 #include "../transform/ojph_transform.h"
@@ -57,105 +58,6 @@ namespace ojph {
 
   namespace local
   {
-
-    //////////////////////////////////////////////////////////////////////////
-    struct bit_write_buf
-    {
-      static const int needed;
-
-      bit_write_buf() { ccl = NULL; avail_bits = 0; tmp = 0; }
-      coded_lists* ccl;
-      int avail_bits;
-      ui64 tmp;
-    };
-
-    //////////////////////////////////////////////////////////////////////////
-    const int bit_write_buf::needed = 512;
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    void bb_expand_buf(bit_write_buf *bbp, mem_elastic_allocator *elastic,
-                       coded_lists*& cur_coded_list)
-    {
-      assert(cur_coded_list == NULL);
-      elastic->get_buffer(bit_write_buf::needed, cur_coded_list);
-      bbp->ccl = cur_coded_list;
-      bbp->tmp = 0;
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    void bb_init(bit_write_buf *bbp, mem_elastic_allocator *elastic,
-                 coded_lists*& cur_coded_list)
-    {
-      bb_expand_buf(bbp, elastic, cur_coded_list);
-      bbp->avail_bits = 8;
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    void bb_put_bit(bit_write_buf *bbp, ui32 bit,
-                    mem_elastic_allocator *elastic,
-                    coded_lists*& cur_coded_list, ui32& ph_bytes)
-    {
-      --bbp->avail_bits;
-      bbp->tmp |= (bit & 1) << bbp->avail_bits;
-      if (bbp->avail_bits <= 0)
-      {
-        bbp->avail_bits = 8 - (bbp->tmp != 0xFF ? 0 : 1);
-        bbp->ccl->buf[bbp->ccl->buf_size - bbp->ccl->avail_size] =
-          (ui8)(bbp->tmp & 0xFF);
-        bbp->tmp = 0;
-        --bbp->ccl->avail_size;
-        if (bbp->ccl->avail_size == 0)
-        {
-          bb_expand_buf(bbp, elastic, cur_coded_list->next_list);
-          cur_coded_list = cur_coded_list->next_list;
-          ph_bytes += bit_write_buf::needed;
-        }
-      }
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    void bb_put_bits(bit_write_buf *bbp, ui32 data, int num_bits,
-                     mem_elastic_allocator *elastic,
-                     coded_lists*& cur_coded_list, ui32& ph_bytes)
-    {
-//      assert(num_bits <= 32);
-      for (int i = num_bits - 1; i >= 0; --i)
-        bb_put_bit(bbp, data >> i, elastic, cur_coded_list, ph_bytes);
-//      while (num_bits) {
-//        int tx_bits = num_bits < bbp->avail_bits ? num_bits : bbp->avail_bits;
-//        bbp->tmp |= (data >> (num_bits - tx_bits)) & ((1 << tx_bits) - 1);
-//        bbp->avail_bits -= tx_bits;
-//        if (bbp->avail_bits <= 0)
-//        {
-//          bbp->avail_bits = 8 - (bbp->tmp != 0xFF ? 0 : 1);
-//          bbp->buf[bbp->buf_size - bbp->avail_size] = (ui8)(bbp->tmp & 0xFF);
-//          bbp->tmp = 0;
-//          --bbp->avail_size;
-//          if (bbp->avail_size == 0)
-//          {
-//            bb_expand_buf(bbp, elastic, cur_coded_list->next_list);
-//            cur_coded_list = cur_coded_list->next_list;
-//            ph_bytes += bit_buffer::needed;
-//          }
-//        }
-//      }
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    void bb_terminate(bit_write_buf *bbp)
-    {
-      if (bbp->avail_bits < 8) //bits have been written
-      {
-        ui8 val = (ui8)(bbp->tmp & 0xFF);
-        bbp->ccl->buf[bbp->ccl->buf_size - bbp->ccl->avail_size] = val;
-        --bbp->ccl->avail_size;
-      }
-    }
 
     //////////////////////////////////////////////////////////////////////////
     struct tag_tree
@@ -422,172 +324,6 @@ namespace ojph {
       }
     }
 
-    //////////////////////////////////////////////////////////////////////////
-    struct bit_read_buf
-    {
-      infile_base *file;
-      ui32 tmp;
-      int avail_bits;
-      bool unstuff;
-      ui32 bytes_left;
-    };
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    void bb_init(bit_read_buf *bbp, ui32 bytes_left, infile_base* file)
-    {
-      bbp->avail_bits = 0;
-      bbp->file = file;
-      bbp->bytes_left = bytes_left;
-      bbp->tmp = 0;
-      bbp->unstuff = false;
-    }
-
-    /////////////////////////////////////////////////////////////////////////////
-    static inline
-    bool bb_read(bit_read_buf *bbp)
-    {
-      if (bbp->bytes_left > 0)
-      {
-        ui32 t = 0;
-        if (bbp->file->read(&t, 1) != 1)
-          throw "error reading from file";
-        bbp->tmp = t;
-        bbp->avail_bits = 8 - bbp->unstuff;
-        bbp->unstuff = (t == 0xFF);
-        --bbp->bytes_left;
-        return true;
-      }
-      else
-      {
-        bbp->tmp = 0;
-        bbp->avail_bits = 8 - bbp->unstuff;
-        bbp->unstuff = false;
-        return false;
-      }
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    bool bb_read_bit(bit_read_buf *bbp, ui32& bit)
-    {
-      bool result = true;
-      if (bbp->avail_bits == 0)
-        result = bb_read(bbp);
-      bit = (bbp->tmp >> --bbp->avail_bits) & 1;
-      return result;
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    bool bb_read_bits(bit_read_buf *bbp, int num_bits, ui32& bits)
-    {
-      assert(num_bits <= 32);
-
-      bits = 0;
-      bool result = true;
-      while (num_bits) {
-        if (bbp->avail_bits == 0)
-          result = bb_read(bbp);
-        int tx_bits = ojph_min(bbp->avail_bits, num_bits);
-        bits <<= tx_bits;
-        bbp->avail_bits -= tx_bits;
-        num_bits -= tx_bits;
-        bits |= (bbp->tmp >> bbp->avail_bits) & ((1 << tx_bits) - 1);
-      }
-      return result;
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    bool bb_read_chunk(bit_read_buf *bbp, ui32 num_bytes,
-                       coded_lists*& cur_coded_list,
-                       mem_elastic_allocator *elastic)
-    {
-      assert(bbp->avail_bits == 0 && bbp->unstuff == false);
-      ui32 bytes = ojph_min(num_bytes, bbp->bytes_left);
-      elastic->get_buffer(bytes + coded_cb_header::prefix_buf_size
-        + coded_cb_header::suffix_buf_size, cur_coded_list);
-      ui32 bytes_read = (ui32)bbp->file->read(
-        cur_coded_list->buf + coded_cb_header::prefix_buf_size, bytes);
-      if (num_bytes > bytes_read)
-        memset(cur_coded_list->buf + coded_cb_header::prefix_buf_size + bytes, 
-          0, num_bytes - bytes_read);
-      bbp->bytes_left -= bytes_read;
-      return bytes_read == bytes;
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    void bb_skip_eph(bit_read_buf *bbp)
-    {
-      if (bbp->bytes_left >= 2)
-      {
-        ui8 marker[2];
-        if (bbp->file->read(marker, 2) != 2)
-          throw "error reading from file";
-        bbp->bytes_left -= 2;
-        if ((int)marker[0] != (EPH >> 8) || (int)marker[1] != (EPH & 0xFF))
-          throw "should find EPH, but found something else";
-      }
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    bool bb_terminate(bit_read_buf *bbp, bool uses_eph)
-    {
-      bool result = true;
-      if (bbp->unstuff)
-        result = bb_read(bbp);
-      assert(bbp->unstuff == false);
-      if (uses_eph)
-        bb_skip_eph(bbp);
-      bbp->tmp = 0;
-      bbp->avail_bits = 0;
-      return result;
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    static inline
-    bool bb_skip_sop(bit_read_buf *bbp)
-    {
-      if (bbp->bytes_left >= 2)
-      {
-        ui8 marker[2];
-        if (bbp->file->read(marker, 2) != 2)
-          throw "error reading from file";
-        if ((int)marker[0] == (SOP >> 8) && (int)marker[1] == (SOP & 0xFF))
-        {
-          bbp->bytes_left -= 2;
-          if (bbp->bytes_left >= 4)
-          {
-            ui16 com_len;
-            if (bbp->file->read(&com_len, 2) != 2)
-              throw "error reading from file";
-            com_len = swap_byte(com_len);
-            if (com_len != 4)
-              throw "something is wrong with SOP length";
-            int result = 
-              bbp->file->seek(com_len - 2, infile_base::OJPH_SEEK_CUR);
-            if (result != 0)
-              throw "error seeking file";
-            bbp->bytes_left -= com_len;
-          }
-          else
-            throw "precinct truncated early";
-          return true;
-        }
-        else
-        {
-          //put the bytes back
-          if (bbp->file->seek(-2, infile_base::OJPH_SEEK_CUR) != 0)
-            throw "error seeking file";
-          return false;
-        }
-      }
-
-      return false;
-    }
 
     //////////////////////////////////////////////////////////////////////////
     void precinct::parse(int tag_tree_size, ui32* lev_idx,
